Rename vector_utils
