+++ src/vector_utils.h
